# Change zstd compression algorithm