--- a/go.mod
+++ b/go.mod
@@ -3,7 +3,6 @@ module github.com/dgraph-io/badger
 go 1.12
 
 require (
-	github.com/DataDog/zstd v1.4.1
 	github.com/cespare/xxhash v1.1.0
 	github.com/cespare/xxhash/v2 v2.1.0 // indirect
 	github.com/dgraph-io/ristretto v0.0.0-20190916120426-cd2835491e0e
@@ -16,6 +15,7 @@ require (
 	github.com/spf13/cobra v0.0.5
 	github.com/stretchr/objx v0.2.0 // indirect
 	github.com/stretchr/testify v1.4.0
+	github.com/valyala/gozstd v1.6.2 // indirect
 	golang.org/x/net v0.0.0-20190620200207-3b0461eec859
 	golang.org/x/sys v0.0.0-20190626221950-04f50cda93cb
 	gopkg.in/check.v1 v1.0.0-20190902080502-41f04d3bba15 // indirect