Currently, we use https://github.com/DataDog/zstd implementation for ZSTD compression and https://github.com/valyala/gozstd seems to be faster than https://github.com/DataDog/zstd . This PR 
proposes use of `valyala/gozstd` instead of `DataDog/zstd`.

```
benchstat zstd.txt gozstd.txt 
name                       old time/op  new time/op  delta
Compression/compress-16     213ms ± 6%   116ms ± 3%  -45.79%  (p=0.000 n=9+10)
Compression/decompress-16  48.5ms ± 3%  45.0ms ± 1%   -7.30%  (p=0.000 n=10+10)
```

The following script was used for benchmarking
```go
var a []byte
func BenchmarkCompression(b *testing.B) {
	opts := Options{BlockSize: 4 * 1024, BloomFalsePositive: 0.01}
	count := uint64(1e6)
	builder := NewTableBuilder(opts)
	// Approx 64 MB table size
	for i := uint64(0); i < count*2; i++ {
		k := fmt.Sprintf("%016x", i)
		v := fmt.Sprintf("%016d", i)
		builder.Add([]byte(k), y.ValueStruct{Value: []byte(v)})
	}
	// Build a dummy table so that we can use the same buffer for benchmarks
	buf := builder.Finish()

	var x []byte
	var err error
	// Used for decompression benchmarks
	// cbuf, err := zstd.Compress(nil, buf) 	// Use this buffer for zstd.
	// require.NoError(b, err)
	cbuf1 := gozstd.Compress(nil, buf)

	fmt.Println("len of buf", humanize.Bytes(uint64(len(buf))))
	b.ResetTimer()
	b.Run("compress", func(b *testing.B) {
		for i := 0; i < b.N; i++ {
			x = gozstd.Compress(nil, buf) // change this for zstd.
		}
		a = x
	})
	b.Run("decompress", func(b *testing.B) {
		for i := 0; i < b.N; i++ {
			x, err = gozstd.Decompress(nil, cbuf1) // change this for zstd.
			if err != nil {
				panic(err)
			}
		}
		a = x
	})
}
```

The following is the benchmark of time taken to build a table in badger (https://github.com/dgraph-io/badger/blob/f50343ff404d8198df6dc83755ec2eab863d5ff2/table/builder_test.go#L116)
```
benchstat master-builder.txt new-builder.txt                       
name        old time/op  new time/op  delta
Builder-16   1.02s ± 1%   0.73s ± 2%  -27.77%  (p=0.000 n=9+10)
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/dgraph-io/badger/1069)
<!-- Reviewable:end -->