@ashish-goswami 2019-10-07
Benchmarks are surely better for `gozstd` than `zstd`. But we should also consider stability and support for both. Just try to find who are users of both libraries(if they are public) and using those in production.