@jarifibrahim 2019-10-14
Benchmarks show LZ4 is faster than zstd and so we don't need this PR anymore.
```
go test -v -run ^$ -bench BenchmarkTableCompression
data size 4056
goos: linux
goarch: amd64
pkg: github.com/dgraph-io/badger/table
BenchmarkTableCompression/Compression/Snappy-16 	   	   	    	  148005	      8103 ns/op
BenchmarkTableCompression/Compression/LZ4-16      	   	   	   	  	  192206	      5589 ns/op
BenchmarkTableCompression/Compression/ZSTD_-_Datadog-16         	   24697	     47433 ns/op
BenchmarkTableCompression/Decompression/Snappy-16               	  308114	      3952 ns/op
BenchmarkTableCompression/Decompression/LZ4-16                  	  379030	      2728 ns/op
BenchmarkTableCompression/Decompression/ZSTD_-_Datadog-16       	  100808	     11431 ns/op
PASS
ok  	github.com/dgraph-io/badger/table	8.708s
```
https://gist.github.com/jarifibrahim/50b596c233aa9d82706aa0df65a546ec used for benchmarking.