@arl 2019-11-30
I know this is closed but there's another reason to switch zstd library, a serious memory corruption bug in DataDog/zstd, that can either to segfaults, lower compression ratio, or nothing depending on your usage of the library.

Please see https://github.com/DataDog/zstd/issues/22

.